consolidate ts types into 1 file for convenience
